third_party/WebKit/LayoutTests/http/tests/devtools/sources/debugger/js-with-inline-stylesheets-expected.txt

Tests that JS sourcemapping for inline scripts followed by inline stylesheets does not break.


Running: testSetBreakpoint
Script source was shown.
Set timer for test function.
Script execution paused.
Call stack:
    0) testFunction (js-with-inline-stylesheets.js:13)
Breakpoint sidebar pane 
js-with-inline-stylesheets.js:13          var x = Math.sqrt(10);
Breakpoint sidebar pane 
No breakpoints
Script execution resumed.
